backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
chromadb
langchain
langchain-community
langchain-core
google-generativeai==0.8.3
python-pptx
python-docx
python-dotenv
sentence-transformers
